feat: convert app dir pages to server components
