+++ packages/dapp/src/components/CredentialDisplay/JsonPanel.tsx
