--- packages/dapp/src/components/VC/tabs/QRTab.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { QueryVCsRequestResult } from '@blockchain-lab-um/masca-types';
-
-interface QRTabProps {
-  vc: QueryVCsRequestResult;
-}
-
-const QRTab = ({ vc }: QRTabProps) => {
-  return <div>QRTab</div>;
-};
-
-export default QRTab;
